Refactor server messages
