+++ src/server/chat/events.rs
@@ -1,0 +1,7 @@
+use crate::prelude::*;
+
+#[derive(Event)]
+pub struct PlayerChatMessageSendEvent {
+    pub client_id: ClientId,
+    pub message: String
+}
