+++ src/server/chat/mod.rs
@@ -1,6 +1,8 @@
 use crate::prelude::*;
 
 pub mod resources;
+pub mod systems;
+pub mod events;
 
 pub struct ChatPlugin;
 
@@ -8,5 +10,7 @@ impl Plugin for ChatPlugin {
     fn build(&self, app: &mut App) {
         info!("Building ChatPlugin");
         app.insert_resource(resources::ChatHistory::new());
+        app.add_systems(Update, chat_systems::handle_network_chat_message_send);
+        app.add_event::<chat_events::PlayerChatMessageSendEvent>();
     }
 }
