+++ src/server/prelude.rs
@@ -38,3 +38,5 @@ pub use crate::terrain::systems as terrain_systems;
 pub use crate::terrain::util as terrain_util;
 
 pub use crate::chat::resources as chat_resources;
+pub use crate::chat::events as chat_events;
+pub use crate::chat::systems as chat_systems;
